backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
mysql-connector-python==8.1.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
bcrypt==4.0.1
celery==5.3.1
redis==4.6.0
